Move default helpfile to separate module.
The default help being in the help module resulted in custom bots not
being able to use their own top-level help since an existing category
description cannot be re-defined. With the default help in a new module
a custom bot can simply not load the help_default module and register a
help file containing the top-level category in its own module.